Gsapdevtools

+++ ani-handoff.js
@@ -5,20 +5,22 @@ const 	Promise = require("bluebird"),
 				rimraf = Promise.promisify(require('rimraf')),
 				FolderZip = require('folder-zip'),
 				imagemin = require('imagemin'),
-				colors = require("colors"),
 				imageminJpegtran = require('imagemin-jpegtran'),
-				imageminPngcrush = require('imagemin-pngcrush'),
-				imageminGiflossy = require('imagemin-giflossy'),
+				imageminPngquant = require('imagemin-pngquant'),
+				colors = require("colors"),
 				$ = require("./utils");
 
 exports.handoff = async function () {
 	const config = JSON.parse(await $.read_path("./ani-conf.json")),
-				first_banner_file = await $.read_path(`./banners/${config.sizes[0]}.html`);
+				first_banner_file = await $.read_path(`./banners/${config.sizes[0]}.html`),
+				preview_path = fs.existsSync("./preview/");
+
 	if (!config || !first_banner_file) return $.handle_error("No project found, please run 'ani init' and 'ani one' to start your project.");
+	if (!preview_path) return $.handle_error("No preview directory found, please run 'ani preview' to generate a project preview first.");
 
 	try {
 		const banner_files = await $.get_files_in("./banners/"),
-					handoff_path = "./" + config.project + "-handoff",
+					handoff_path = "./preview/" + config.project + "-handoff",
 					handoff_path_exists = $.read_path(handoff_path);
 
 		if (handoff_path_exists) await rimraf(handoff_path);
@@ -29,13 +31,18 @@ exports.handoff = async function () {
 			let static_files = await $.read_dir("./assets/statics");
 			if (static_files) {
 				await fs.mkdirAsync(`${handoff_path}/statics`);
-				await imagemin([`./assets/statics/*.{jpg,png,gif}`], `${handoff_path}/statics`, {
-					plugins: [
-						imageminJpegtran(),
-						imageminPngcrush(),
-						imageminGiflossy({lossy: 0}),
-					]
-				});
+				await imagemin([`./assets/statics/*.{jpg,png,gif}`],
+					{
+						destination: `${handoff_path}/statics`,
+						plugins: [
+							imageminJpegtran(),
+							imageminPngquant({
+								quality: [0.6, 0.8]
+							})
+						]
+					}
+				);
+
 				// Rename statics with project name
 				await fs.readdir(`${handoff_path}/statics`, (err, files) => {
 					for (const file of files) {
@@ -49,11 +56,11 @@ exports.handoff = async function () {
 				$.handle_notice("No static files found. As such, none will be added to the handoff.");
 			}
 		} catch (e) {
-			$.handle_notice("No static files found. As such, none will be added to the handoff.");
+			$.handle_error("Problem generating static files.");
 		}
 		// end statics
 
-		for (let vendor_name in config.vendors)	{
+		for (let vendor_name in config.vendors) {
 			const vendor_path = `${handoff_path}/${vendor_name}`;
 			await fs.mkdirAsync(vendor_path);
 
@@ -69,10 +76,9 @@ exports.handoff = async function () {
 					await zipFolder(banner.path, {excludeParentFolder: true});
 					await writeToFile(`${banner.path}.zip`);
 					await fs.rename(`${banner.path}.zip`, `${vendor_path}/${config.project}-${banner.size}.zip`, (err) => {
-							if (err) throw err
+							if (err) $.handle_error( err )
 						});
 					await rimraf(banner.path)
-
 				} catch (e) {
 					$.handle_error(`Failed to gather assets for or zip ${banner_info.filename}`);
 				}
